Создал два модуля data.js и util.js
